Implement delete comment.

+++ controllers/api/comment_routes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { DataTypes } = require('sequelize');
-const { Comment, Post } = require('../../models');
+const { Comment, Post, User } = require('../../models');
 const { withAuthAPI } = require('../../utils/auth');
 
 // Create 1 comment
@@ -51,14 +51,25 @@ router.put('/:id', withAuthAPI, async (req, res) => {
 // Delete 1 comment
 router.delete('/:id', withAuthAPI, async (req, res) => {
   try {
-    const delete_result = await Comment.destroy({
-      where: {id: req.params.id}
+    // Make sure user is the author of the comment
+    const comment_data = await Comment.findByPk(req.params.id, {
+      include: [{model: User}]
     });
-
-    if (delete_result) {
+    if (!comment_data) {
       res.status(404).json({message: 'Comment with given ID not found!'});
       return;
     }
+    const comment = comment_data.get({plain:true});
+    console.log(comment.user.id);
+    console.log(req.session.user_id);
+    if (req.session.user_id != comment.user.id) {
+      res.status(401).json({message: 'Your user id does not match the author of the comment you are trying to delete.'});
+      return;
+    }
+
+    const delete_result = await Comment.destroy({
+      where: {id: req.params.id}
+    });
 
     res.status(200).json({message: 'Comment deleted successfully'});
